+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased Changes]
 - [Report diagnostics in `Type::from_ast`](https://github.com/amp-lang/ampc/issues/5) (report invalid types)
+- Report type mismatches in `const` declarations
 
 ### [`snapshot-02-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-02-06-2023-a)
 (initial snapshot)